feat: NLU V4.2 + Multi-Agent Runtime V1/V2 + Trace P3
    - NLU Pipeline: 1 LLM call, alias resolver, intent reranker, route hint, smart merge
    - Multi-Agent V1: multi-intent serial execution, agent-turn SSE, source tracking
    - Multi-Agent V2: WorkflowMatcher, TaskExecutor, ResultAggregator, TokenBudget
    - Trace P3: TraceTimelineView in chat, real-time trace timeline
    - Frontend: agent avatars, clarification events, history source adaptation
    - Docs: NLU V4.2 design, WIKI/README/FEATURES updated"

+++ src/main/java/com/yupi/yuaiagent/mcp/McpAuditLog.java
@@ -72,11 +72,11 @@ public class McpAuditLog {
      * Get average latency for a tool.
      */
     public long getAvgLatency(String serverName, String toolName) {
-        return entries.stream()
+        return (long) entries.stream()
                 .filter(e -> e.serverName.equals(serverName) && e.toolName.equals(toolName))
                 .mapToLong(e -> e.latencyMs)
                 .average()
-                .orElse(-1.0).longValue();
+                .orElse(-1.0);
     }
 
     /**
